some hardening for message format and other improvements
We need to pull the content out of a RMQ message, and parse it (it
should be JSON). I've added a sample of a successfully-fetched build
(which itself was a failure). This at least proves we have a working
pipeline from Jenkins to RabbitMQ to our app, as well as an API
call back to Jenkins!